Refactor the signing code to make it reusable
